requirements.txt
discord==1.7.3
openai==0.27.0
python-dotenv
